[bash] Limit memory for processes to 80% of total RAM
Sometimes there is a rogue process which thinks it has to keep allocating,
eventually forcing the system to start swapping.  If you have a huge swap (to
enable suspend to disk), it takes ages for the swap to be full and the process
being killed.  Therefore I just limit everything I start from the shell to 80%
RAM.