[nix] Just use what's on the channel
Previously I had pinned the version of nixpkgs to a particular channel
to override the system channel.  However, in the meantime I have
switched to declarative channels for the system configuration in NixOS
and elsewhere I have to manually handle channels anyway so there is
little advantage from this strategy anymore.  By just using the system
channel I also avoid having the user environment and the system diverge.